cj50.h: failed attempt at `EQUAL`
Can't build on top of borrowing functions, except to always consume.

But the aim was to allow to borrow from function returns, and that
won't work if also mixed with variables that should not be
consumed. Because it's about borrowing, remember. It's just that
nobody drops the function returns, then.

Auto drop would really be the right solution...