Result.h: add drop_Result_$T__$E
I had this a long time ago, then removed it because the normal
approach is to extract the contained value.

But, this is to be left to the user, and offering drop is fine! Just
don't call drop if the value was moved out already. And equal_move_*
definitely needs it.